+++ ts-client/src/dlmm/types/index.ts
@@ -114,6 +114,7 @@ export type CompressedBinDepositAmounts = CompressedBinDepositAmount[];
 export type ResizeSideEnum = IdlTypes<LbClmm>["resizeSide"];
 export type ExtendedPositionBinData = IdlTypes<LbClmm>["positionBinData"];
 
+<<<<<<< HEAD
 export type LimitOrderBinData = IdlTypes<LbClmm>["limitOrderBinData"];
 export type PlaceLimitOrderParams = IdlTypes<LbClmm>["placeLimitOrderParams"];
 
@@ -121,6 +122,9 @@ export interface ParsedLimitOrderWithPubkey {
   publicKey: PublicKey;
   limitOrderData: ParsedLimitOrder;
 }
+=======
+export type Oracle = IdlTypes<LbClmm>["oracle"];
+>>>>>>> origin/main
 
 export interface LbPosition {
   publicKey: PublicKey;
@@ -165,6 +169,8 @@ export enum PositionVersion {
 export enum PairType {
   Permissionless,
   Permissioned,
+  CustomizablePermissionless,
+  PermissionlessV2,
 }
 
 export enum ShrinkMode {
@@ -666,6 +672,7 @@ export interface GetPositionsOpt {
    */
   isParallelExecution?: boolean;
 }
+<<<<<<< HEAD
 
 export enum PositionPermission {
   ClaimFee,
@@ -676,3 +683,5 @@ export enum LimitOrderStatus {
   PartialFilled,
   Fulfilled,
 }
+=======
+>>>>>>> origin/main